[back] refactor: img_mgr is a separated module
